OptimalPath/requirements.txt
networkx==3.1
matplotlib==3.7.1
ortools==9.6.2534
osmnx==1.3.0
streamlit>=1.22.0,<1.30.0
protobuf>=4.21.12
numpy==1.24.3
pandas==2.0.3
folium==0.14.0
streamlit-folium==0.15.0
streamlit-js-eval>=0.1.5
python-docx==1.0.1
PyPDF2==3.0.1
openpyxl>=3.1.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2
geopy==2.4.0
openrouteservice==2.3.1
kivy==2.2.1
kivymd==1.1.1